Fix: tag can be undefined

+++ backend/src/meta/meta.service.ts
@@ -56,7 +56,7 @@ export class MetaService {
                 data: await this.buildMetaResponsaveis(meta.id, op, cp),
             });
 
-            if (tags.length > 0)
+            if (typeof(tags) != 'undefined' && tags.length > 0)
                 await prisma.metaTag.createMany({
                     data: await this.buildTags(meta.id, tags)
                 });
@@ -300,7 +300,7 @@ export class MetaService {
                 }
             }
 
-            if (tags) {
+            if (typeof(tags) != 'undefined' && tags.length > 0) {
                 await prisma.metaTag.deleteMany({ where: { meta_id: id } });
                 await prisma.metaTag.createMany({ data: await this.buildTags(meta.id, tags) });
             }
